Make the first build success.
Changing some fields from 32bit to 64bit by uintptr_t replacing int32_t. 
memset() with void * for the target .
Adding some objects to "clean" target to make it easier to do clean.
Adding .gitignore file to ignore some project related files.

+++ src/vmobjects/VMFreeObject.cpp
@@ -33,7 +33,7 @@ VMFreeObject::VMFreeObject() : VMObject(0) {
 }
 
 void VMFreeObject::SetNext(VMFreeObject* next) {
-    this->hash = (int32_t) next;
+    this->hash = (uintptr_t) next;
 }
 
 VMFreeObject* VMFreeObject::GetNext() {
@@ -41,7 +41,7 @@ VMFreeObject* VMFreeObject::GetNext() {
 }
 
 void VMFreeObject::SetPrevious(VMFreeObject* prev) {
-    this->numberOfFields = (int32_t)prev;
+    this->numberOfFields = (uintptr_t)prev;
 }
 
 VMFreeObject* VMFreeObject::GetPrevious() {
